+++ tests/tests_forums.py
@@ -3,8 +3,8 @@ import unittest
 
 import tibiapy
 from tests.tests_tibiapy import TestCommons
-from tibiapy import CMPostArchive, ForumAnnouncement, ForumBoard, ForumThread, InvalidContent, LastPost, ListedBoard, \
-    ListedThread, ThreadStatus
+from tibiapy import CMPostArchive, ForumAnnouncement, ForumBoard, ForumThread, InvalidContent, LastPost, BoardEntry, \
+    ThreadEntry, ThreadStatus
 
 FILE_WORLD_BOARDS = "forums/tibiacom_section.txt"
 FILE_SECTION_EMPTY_BOARD = "forums/tibiacom_section_empty_board.txt"
@@ -28,12 +28,12 @@ class TestForum(TestCommons, unittest.TestCase):
     def test_listed_board_list_from_content_world_boards(self):
         content = self.load_resource(FILE_WORLD_BOARDS)
 
-        boards = ListedBoard.list_from_content(content)
+        boards = BoardEntry.list_from_content(content)
 
         self.assertEqual(77, len(boards))
         for i, board in enumerate(boards):
             with self.subTest(i=i):
-                self.assertIsInstance(board, ListedBoard)
+                self.assertIsInstance(board, BoardEntry)
                 self.assertIsNotNone(board.name)
                 self.assertGreater(board.board_id, 0)
                 self.assertGreater(board.posts, 0)
@@ -43,12 +43,12 @@ class TestForum(TestCommons, unittest.TestCase):
     def test_listed_board_list_from_content_empty_board(self):
         content = self.load_resource(FILE_SECTION_EMPTY_BOARD)
 
-        boards = ListedBoard.list_from_content(content)
+        boards = BoardEntry.list_from_content(content)
 
         self.assertEqual(77, len(boards))
         for i, board in enumerate(boards):
             with self.subTest(i=i):
-                self.assertIsInstance(board, ListedBoard)
+                self.assertIsInstance(board, BoardEntry)
                 self.assertIsNotNone(board.name)
                 self.assertGreater(board.board_id, 0)
                 self.assertGreaterEqual(board.posts, 0)
@@ -58,7 +58,7 @@ class TestForum(TestCommons, unittest.TestCase):
     def test_listed_board_list_from_content_empty_section(self):
         content = self.load_resource(FILE_SECTION_EMPTY)
 
-        boards = ListedBoard.list_from_content(content)
+        boards = BoardEntry.list_from_content(content)
 
         self.assertIsInstance(boards, list)
 
@@ -66,7 +66,7 @@ class TestForum(TestCommons, unittest.TestCase):
         content = self.load_resource(self.FILE_UNRELATED_SECTION)
 
         with self.assertRaises(InvalidContent):
-            ListedBoard.list_from_content(content)
+            BoardEntry.list_from_content(content)
 
     def test_forum_board_from_content(self):
         content = self.load_resource(FILE_BOARD_THREAD_LIST)
@@ -82,10 +82,10 @@ class TestForum(TestCommons, unittest.TestCase):
         self.assertEqual(30, len(board.threads))
         self.assertIsNotNone(board.url)
         self.assertIsNotNone(board.next_page_url)
-        self.assertEqual(board.next_page_url, ListedBoard.get_url(board.board_id, board.page+1, board.age))
+        self.assertEqual(board.next_page_url, BoardEntry.get_url(board.board_id, board.page + 1, board.age))
         for i, thread in enumerate(board.threads):
             with self.subTest(i=i):
-                self.assertIsInstance(thread, ListedThread)
+                self.assertIsInstance(thread, ThreadEntry)
                 self.assertIsNotNone(thread.title)
                 self.assertGreater(thread.thread_id, 0)
                 self.assertGreaterEqual(thread.views, 0)
@@ -97,10 +97,10 @@ class TestForum(TestCommons, unittest.TestCase):
             board.get_page_url(-1)
 
     def test_board_section_urls(self):
-        self.assertIsNotNone(ListedBoard.get_community_boards_url())
-        self.assertIsNotNone(ListedBoard.get_support_boards_url())
-        self.assertIsNotNone(ListedBoard.get_world_boards_url())
-        self.assertIsNotNone(ListedBoard.get_trade_boards_url())
+        self.assertIsNotNone(BoardEntry.get_community_boards_url())
+        self.assertIsNotNone(BoardEntry.get_support_boards_url())
+        self.assertIsNotNone(BoardEntry.get_world_boards_url())
+        self.assertIsNotNone(BoardEntry.get_trade_boards_url())
 
     def test_forum_board_from_content_empty_threads(self):
         content = self.load_resource(FILE_BOARD_EMPTY_THREAD_LIST)